グローバル変数の範囲を縮小することを念頭にファイルの分割を実施
