+++ server/client/bot.go
@@ -1,9 +1,9 @@
-package handler
+package client
 
 import (
-	"traQ-gazer/api"
-	"traQ-gazer/model"
 	"net/http"
+	"traQ-gazer/api"
+	"traQ-gazer/db"
 
 	"github.com/labstack/echo/v4"
 )
@@ -13,7 +13,7 @@ import (
 func (s Server) PutWords(ctx echo.Context) error {
 	data := &api.WordBotSetting{}
 	err := ctx.Bind(data)
-	
+
 	if err != nil {
 		// 正常でないためステータスコード 400 "Invalid Input"
 		return echo.NewHTTPError(http.StatusBadRequest, err)
@@ -26,7 +26,7 @@ func (s Server) PutWords(ctx echo.Context) error {
 		return echo.NewHTTPError(http.StatusBadRequest, err)
 	}
 
-	exist, err := model.ExistWord(data.Word, userId)
+	exist, err := db.ExistWord(data.Word, userId)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err)
 	}
@@ -35,11 +35,11 @@ func (s Server) PutWords(ctx echo.Context) error {
 		return echo.NewHTTPError(http.StatusNotFound, "Not Found")
 	}
 
-	err = model.ChengeBotNotification(data.Word, data.IncludeBot, userId)
+	err = db.ChengeBotNotification(data.Word, data.IncludeBot, userId)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err)
 	}
-	
+
 	return ctx.JSON(http.StatusOK, "Successful Change")
 }
 
@@ -48,7 +48,7 @@ func (s Server) PutWords(ctx echo.Context) error {
 func (s Server) PostWordsBot(ctx echo.Context) error {
 	data := &api.Bot{}
 	err := ctx.Bind(data)
-	
+
 	if err != nil {
 		// 正常でないためステータスコード 400 "Invalid Input"
 		return echo.NewHTTPError(http.StatusBadRequest, err)
@@ -61,7 +61,7 @@ func (s Server) PostWordsBot(ctx echo.Context) error {
 		return echo.NewHTTPError(http.StatusBadRequest, err)
 	}
 
-	err = model.ChangeAllBotNotification(data.IncludeBot, userId)
+	err = db.ChangeAllBotNotification(data.IncludeBot, userId)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err)
 	}
