+++ server/client/list.go
@@ -1,9 +1,10 @@
-package handler
+package client
 
 import (
-	"traQ-gazer/api"
-	"traQ-gazer/model"
 	"net/http"
+	"traQ-gazer/api"
+	"traQ-gazer/db"
+	"traQ-gazer/model"
 
 	"github.com/labstack/echo/v4"
 )
@@ -18,7 +19,7 @@ func (s Server) GetListUserMe(ctx echo.Context) error {
 		return echo.NewHTTPError(http.StatusBadRequest, err)
 	}
 
-	wordsListMode, err := model.GetListUserUserId(userId)
+	wordsListMode, err := db.GetListUserUserId(userId)
 	if err != nil {
 		return err
 	}
@@ -29,7 +30,7 @@ func (s Server) GetListUserMe(ctx echo.Context) error {
 // あるuserのwordたち
 // (GET /list/user/{userId})
 func (s Server) GetListUserUserId(ctx echo.Context, userId string) error {
-	wordsListMode, err := model.GetListUserUserId(userId)
+	wordsListMode, err := db.GetListUserUserId(userId)
 	if err != nil {
 		return err
 	}
@@ -66,7 +67,7 @@ func ConvertWordList(models model.WordsList) api.WordsList {
 // ある単語を見ているuserたち
 // (GET /list/word/{word})
 func (s Server) GetListWordWord(ctx echo.Context, word string) error {
-	usersListMode, err := model.GetListWordWord(word)
+	usersListMode, err := db.GetListWordWord(word)
 	if err != nil {
 		return echo.NewHTTPError(500, err.Error())
 	}
