+++ server/client/session.go
@@ -1,4 +1,4 @@
-package handler
+package client
 
 import (
 	"errors"
