+++ server/db/me.go
@@ -1,4 +1,4 @@
-package model
+package db
 
 func ChengeMeNotification(word string, includeMe bool, userId string) error {
 	_, err := db.Exec("UPDATE words SET me_notification = ? WHERE word = ? AND trap_id = ?", includeMe, word, userId)
