+++ server/traqmessage/collect.go
@@ -6,6 +6,8 @@ import (
 	"strings"
 	"sync"
 	"time"
+	"traQ-gazer/bot"
+	"traQ-gazer/db"
 	"traQ-gazer/model"
 
 	"github.com/traPtitech/go-traq"
@@ -30,7 +32,7 @@ func (m *MessagePoller) Run() {
 
 	const pollingInterval = time.Minute * 3
 
-	lastCheckpoint, err := model.GetPollingFrom()
+	lastCheckpoint, err := db.GetPollingFrom()
 	if err != nil {
 		slog.Error(fmt.Sprintf("Failed to get pollinginfo: %v", err))
 		lastCheckpoint = time.Now()
@@ -80,7 +82,7 @@ func (m *MessagePoller) Run() {
 
 		slog.Info(fmt.Sprintf("%d messages collected totally", collectedMessageCount))
 
-		err := model.RecordPollingTime(lastCheckpoint)
+		err := db.RecordPollingTime(lastCheckpoint)
 		if err != nil {
 			slog.Error(fmt.Sprintf("Failed to recording lastCheckpoint: %v", err))
 		}
@@ -114,7 +116,7 @@ func (m *messageProcessor) process(messages []traq.Message) {
 		slog.Error(fmt.Sprintf("Failed to convert messages: %v", err))
 		return
 	}
-	notifyInfoList, err := model.FindMatchingWords(messageList)
+	notifyInfoList, err := bot.FindMatchingWords(messageList)
 	if err != nil {
 		slog.Error(fmt.Sprintf("Failed to process messages: %v", err))
 		return
@@ -144,13 +146,13 @@ func genNotifyMessageContent(citeMessageId string, words ...string) string {
 }
 
 func sendMessage(notifyTargetTraqUUID string, messageContent string) error {
-	if model.ACCESS_TOKEN == "" {
+	if db.ACCESS_TOKEN == "" {
 		slog.Info("Skip sendMessage")
 		return nil
 	}
 
 	client := traq.NewAPIClient(traq.NewConfiguration())
-	auth := context.WithValue(context.Background(), traq.ContextAccessToken, model.ACCESS_TOKEN)
+	auth := context.WithValue(context.Background(), traq.ContextAccessToken, db.ACCESS_TOKEN)
 	_, _, err := client.UserApi.PostDirectMessage(auth, notifyTargetTraqUUID).PostMessageRequest(traq.PostMessageRequest{
 		Content: messageContent,
 	}).Execute()
@@ -162,13 +164,13 @@ func sendMessage(notifyTargetTraqUUID string, messageContent string) error {
 }
 
 func collectMessages(from time.Time, to time.Time, page int) (*[]traq.Message, bool, error) {
-	if model.ACCESS_TOKEN == "" {
+	if db.ACCESS_TOKEN == "" {
 		slog.Info("Skip collectMessage")
 		return &[]traq.Message{}, false, nil
 	}
 
 	client := traq.NewAPIClient(traq.NewConfiguration())
-	auth := context.WithValue(context.Background(), traq.ContextAccessToken, model.ACCESS_TOKEN)
+	auth := context.WithValue(context.Background(), traq.ContextAccessToken, db.ACCESS_TOKEN)
 
 	// 1度での取得上限は100まで　それ以上はoffsetを使うこと
 	// https://github.com/traPtitech/traQ/blob/47ed2cf94b2209c8444533326dee2a588936d5e0/service/search/engine.go#L51
